third_party/blink/web_tests/fast/css/invalidation/content-attr-expected.txt

No subtree recalc when changing attribute used in generated content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
